requirements.txt
Flask
qrcode[pil]
Pillow>=10.0.0
